requirements.txt
# Web Framework y APIs
flask==3.0.0
flask-cors==4.0.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Data Processing
pandas==2.1.0
pyarrow==13.0.0
duckdb==0.9.0
openpyxl==3.1.2

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.9

# LLM and AI

# Testing
pytest==7.4.0
pytest-cov==4.1.0